requirements.txt
# Core backend dependencies (sync FastAPI app)
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==1.3.1
sqlalchemy==1.4.52
# Use PyMySQL 0.9.3 to keep compatibility with aiomysql 0.0.22 which requires PyMySQL>=0.9,<0.10
PyMySQL==0.9.3
aiomysql==0.0.22
redis==4.5.5
python-dotenv==1.0.0

# Utilities / testing
pytest==7.4.0
httpx==0.24.1

# Dev / lint (optional)
ruff==0.1.2
